+++ frontend/src/App.js
@@ -1,5 +1,5 @@
-import Menu from './components/Menu';
-import Login from './components/Login';
+import Menu from './pages/Menu';
+import Login from './pages/Login';
 export default function App() {
   return (
     <>
