+++ frontend/src/pages/Login.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import Button from './Button';
+import Button from '../components/Button';
 
 export default function Login() {
   // Testing useEffect not relevant to this project
